requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.0
requests==2.31.0
feedparser==6.0.10
twilio==8.10.0
python-dotenv==1.0.0
beautifulsoup4==4.12.0
APScheduler==3.10.0
